Massive Amount of Changes
A ton of debugging and additions went into this which is the first
"tenable" alpha version of typySim. This means that the code does a
basic MC growth in a way that (hopefully) will scale. Lot's more changes
and improvements to come. At this point, the performance of the chain
growth algorithm is worrisome and some improvements will be sought.